third_party/blink/web_tests/http/tests/security/postMessage/target-origin-same-site-expected.txt

CONSOLE WARNING: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://127.0.0.1:9090') does not match the recipient window's origin ('http://127.0.0.1:8000').
CONSOLE WARNING: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://127.0.0.1') does not match the recipient window's origin ('http://127.0.0.1:8000').
CONSOLE WARNING: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://127.0.0.1') does not match the recipient window's origin ('http://127.0.0.1:8000').
CONSOLE WARNING: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://127.0.0.1:8000') does not match the recipient window's origin ('http://127.0.0.1:8000').
CONSOLE WARNING: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://www.example.com') does not match the recipient window's origin ('http://127.0.0.1:8000').
window.location.href = http://127.0.0.1:8000/security/postMessage/target-origin-same-site.html
waiting...
Error sending message to asdf:. SyntaxError: Failed to execute 'postMessage' on 'Window': Invalid target origin 'asdf:' in a call to 'postMessage'.
Error sending message to . SyntaxError: Failed to execute 'postMessage' on 'Window': Invalid target origin '' in a call to 'postMessage'.
Error sending message to asdf. SyntaxError: Failed to execute 'postMessage' on 'Window': Invalid target origin 'asdf' in a call to 'postMessage'.
Error sending message to /tmp/foo. SyntaxError: Failed to execute 'postMessage' on 'Window': Invalid target origin '/tmp/foo' in a call to 'postMessage'.
Error sending message to //localhost. SyntaxError: Failed to execute 'postMessage' on 'Window': Invalid target origin '//localhost' in a call to 'postMessage'.
Error sending message to http:. SyntaxError: Failed to execute 'postMessage' on 'Window': Invalid target origin 'http:' in a call to 'postMessage'.
Received message: data="Received message: data="Trying origin=http://127.0.0.1:8000" origin="http://127.0.0.1:8000"" origin="http://127.0.0.1:8000"
Received message: data="Received message: data="Trying origin=http://127.0.0.1:8000/" origin="http://127.0.0.1:8000"" origin="http://127.0.0.1:8000"
Received message: data="Received message: data="Trying origin=http://127.0.0.1:8000/foo" origin="http://127.0.0.1:8000"" origin="http://127.0.0.1:8000"
Received message: data="Received message: data="Trying origin=http://127.0.0.1:8000/foo?bar" origin="http://127.0.0.1:8000"" origin="http://127.0.0.1:8000"
Received message: data="Received message: data="Trying origin=http://127.0.0.1:8000/foo?bar#baz" origin="http://127.0.0.1:8000"" origin="http://127.0.0.1:8000"
Received message: data="Received message: data="Trying origin=http://user:pass@127.0.0.1:8000/foo?bar#baz" origin="http://127.0.0.1:8000"" origin="http://127.0.0.1:8000"
